M2S_cellToString added; M2S_calculateResiduals with number neighbours = 1 uses the actual feature and no neighbours
Added M2S_cellToString, which is a help function in M2S_tableCellToString, the function to format a table into string or double.
M2S_calculateResiduals was updated. It may or not calculate robust loess to define the inter-dataset shifts, and also when the number of neighbours is 1, then each of the features is selected, rather than their neighbours.